Handle vlan_tci1
This field appears in the "Flow" section of ofproto/trace output. It
is undocumented in ovs-fields(7) but appears to be very much like
the vlan_tci. This commit adds it as a field so that ofproto/trace
output that includes it can be parsed.

It looks like this field is used where dl_vlan (12 bits) and
dl_vlan_pcp (3 bits) also appear. It could be that this field is meant
to convey the 1 bit of the 16 bit vlan tci field that is not already
conveyed by the other two whereas the vlan_tci field is meant to
convey all 16 bits at once.